Move types and events into single folder
