+++ src/features/blacksmith/components/CraftingItems.tsx
@@ -9,9 +9,9 @@ import { OuterPanel } from "components/ui/Panel";
 import { Button } from "components/ui/Button";
 
 import { Context } from "features/game/GameProvider";
-import { ITEM_DETAILS } from "features/game/lib/items";
-import { Craftable } from "features/game/events/craft";
-import { InventoryItemName } from "features/game/lib/types";
+import { ITEM_DETAILS } from "features/game/types/items";
+import { Craftable } from "features/game/types/craftables";
+import { InventoryItemName } from "features/game/types/game";
 
 interface Props {
   items: Partial<Record<InventoryItemName, Craftable>>;
