+++ src/features/game/events/harvest.ts
@@ -1,5 +1,5 @@
-import { CROPS } from "features/crops/lib/crops";
-import { GameState } from "../lib/types";
+import { GameState } from "../types/game";
+import { CROPS } from "../types/crops";
 
 export type HarvestAction = {
   type: "item.harvested";
